Add failing test for computation with "non-proportional" unit mapping
Computation fails with a "non-proportional" unit mapping, e.g.
(.kilometersPerHour, .hours, .centimeters), because 1 km/h * 1 h != 1 cm.
This is because the calculation currently is performed in the desired
unit mapping, i.e. multiplying km/h * h and exptecing the result to be in
cm. The calculation must be performed in the default unit mapping and
then converted to the desired unit mapping.