Add UnitSquare protocol
For relations where one Dimension is the square of another one, such as
UnitArea = UnitLength * UnitLength.

We need a separate protocol for this because using UnitProduct leads to
"ambiguous use of operator '*'" errors because the compiler sees two
matching implementations: (Factor1 * Factor2) and (Factor2 * Factor1)
are identical when Factor1 == Factor2.

It would be nice if it were possible to add a generic constraint like
where Factor1 != Factor2 to one of the overloads, but that is not
suported in Swift.